backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
mistralai>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.1
google-generativeai>=0.3.2
chromadb>=0.4.22
httpx>=0.28.0
python-multipart>=0.0.6
